reorder some stuff
