+++ crates/keepac-cli/src/main.rs
@@ -45,19 +45,19 @@ type SubcommandResult = Result<(), KeepacCliError>;
 fn run(cli: Cli, path: &Path) -> SubcommandResult {
     let command = cli.command.unwrap_or(Command::Show {});
     match command {
-        Command::Add {} => todo!(),
-        Command::Change {} => todo!(),
-        Command::Deprecate {} => todo!(),
+        /* change */ Command::Add {} => todo!(),
+        /* change */ Command::Change {} => todo!(),
+        /* change */ Command::Deprecate {} => todo!(),
         Command::Diff {} => todo!(),
         Command::Edit {} => commands::edit(path),
         Command::Find {} => commands::find(path),
-        Command::Fix {} => todo!(),
+        /* change */ Command::Fix {} => todo!(),
         Command::Init {} => commands::init(path),
         Command::Insert {} => todo!(),
         Command::Release {} => todo!(),
-        Command::Remove {} => todo!(),
+        /* change */ Command::Remove {} => todo!(),
         Command::Search {} => todo!(),
-        Command::Secure {} => todo!(),
+        /* change */ Command::Secure {} => todo!(),
         Command::Show {} => commands::show(path),
         Command::Versions {} => commands::versions(path),
         Command::Yank {} => todo!(),
